requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
fastapi==0.119.0
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
motor==3.7.1
packaging==25.0
pydantic==2.12.0
pydantic_core==2.41.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
Werkzeug==3.1.3
